requirements.txt
Django==3.2.2
django-jazzmin==2.4.8
requests==2.25.1
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary>=2.8

